Feat jinho

+++ src/main/java/com/sparta/springtrello/domain/comment/service/CommentService.java
@@ -1,14 +1,19 @@
 package com.sparta.springtrello.domain.comment.service;
 
+
 import com.sparta.springtrello.domain.comment.dto.request.CommentEditRequestDto;
 import com.sparta.springtrello.domain.comment.dto.request.CommentSaveRequestDto;
 import com.sparta.springtrello.domain.comment.dto.response.CommentEditResponseDto;
 import com.sparta.springtrello.domain.comment.dto.response.CommentSaveResponseDto;
 import com.sparta.springtrello.domain.comment.entity.Comment;
 import com.sparta.springtrello.domain.comment.repository.CommentRepository;
+import com.sparta.springtrello.domain.member.entity.Member;
+import com.sparta.springtrello.domain.member.entity.MemberRole;
 import com.sparta.springtrello.domain.member.repository.MemberRepository;
+import com.sparta.springtrello.domain.ticket.entity.Ticket;
+import com.sparta.springtrello.domain.ticket.repository.TicketRepository;
 import com.sparta.springtrello.domain.user.dto.AuthUser;
-import com.sparta.springtrello.domain.user.repository.UserRepository;
+import com.sparta.springtrello.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,50 +24,45 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
     private final CommentRepository commentRepository;
-    private final UserRepository userRepository;
     private final MemberRepository memberRepository;
+    private final UserService userService;
+    private final TicketRepository ticketRepository;
 
     public CommentSaveResponseDto saveComment(AuthUser authUser, Long cardId, CommentSaveRequestDto commentSaveRequestDto) {
 
         Long userId = authUser.getId();
+        userService.checkUser(userId);
 
-        if (userRepository.findById(userId).isPresent()) {
-            throw new IllegalArgumentException("유저 없음");
+        Ticket ticket = ticketRepository.findById(cardId).orElseThrow(() ->
+                new IllegalArgumentException("card없음"));
+
+        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
+                new IllegalArgumentException("멤버 등록안됨"));
+
+
+        if (!member.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("CREATOR만 수정, 삭제할 수 있음.");
         }
 
-//        Card card = cardRepository.findById(cardId).orElseThrow(() ->
-//                new IllegalArgumentException("card없음"));
-//
-//        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
-//                new IllegalArgumentException("멤버 등록안됨"));
-//
-//        if (memberRepository.findRoleByUserId(userId).equals(Reader)) {
-//            throw new IllegalArgumentException("쓰기/수정 권한 없음");
-//        }
-
-        Comment comment = commentRepository.save(new Comment(commentSaveRequestDto));
-
+        Comment comment = commentRepository.save(new Comment(commentSaveRequestDto, ticket, member));
 
         return new CommentSaveResponseDto(comment);
     }
 
-    public CommentEditResponseDto editComment(AuthUser authUser, Long id, CommentEditRequestDto commentEditRequestDto) {
+    public CommentEditResponseDto editComment(AuthUser authUser, Long id, Long cardId, CommentEditRequestDto commentEditRequestDto) {
 
         Long userId = authUser.getId();
+        userService.checkUser(userId);
 
-        if (userRepository.findById(userId).isPresent()) {
-            throw new IllegalArgumentException("유저 없음");
+        Ticket ticket = ticketRepository.findById(cardId).orElseThrow(() ->
+                new IllegalArgumentException("card없음"));
+
+        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
+                new IllegalArgumentException("멤버 등록안됨"));
+
+        if (!member.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("CREATOR만 수정, 삭제할 수 있음.");
         }
-
-//        Card card = cardRepository.findById(cardId).orElseThrow(() ->
-//                new IllegalArgumentException("card없음"));
-//
-//        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
-//                new IllegalArgumentException("멤버 등록안됨"));
-//
-//        if (memberRepository.findRoleByUserId(userId).equals(Reader)) {
-//            throw new IllegalArgumentException("쓰기/수정 권한 없음");
-//        }
 
         Comment comment = commentRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("댓글 없음"));
@@ -73,23 +73,21 @@ public class CommentService {
         return new CommentEditResponseDto(comment);
     }
 
-    public void deleteComment(Long id, AuthUser authUser) {
+    public void deleteComment(Long id, Long cardId, AuthUser authUser) {
 
-//        if (userRepository.findById(userId).isPresent()) {
-//            throw new IllegalArgumentException("유저 없음");
-//        }
-//
-//        Card card = cardRepository.findById(cardId).orElseThrow(() ->
-//                new IllegalArgumentException("card없음"));
-//
-//        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
-//                new IllegalArgumentException("멤버 등록안됨"));
-//
-//        if (memberRepository.findRoleByUserId(userId).equals(Reader)) {
-//            throw new IllegalArgumentException("쓰기/수정 권한 없음");
-//        }
+        Long userId = authUser.getId();
+        userService.checkUser(userId);
+
+        Ticket ticket = ticketRepository.findById(cardId).orElseThrow(() ->
+                new IllegalArgumentException("card없음"));
+
+        Member member = memberRepository.findByUserId(userId).orElseThrow(() ->
+                new IllegalArgumentException("멤버 등록안됨"));
+
+        if (!member.getMemberRole().equals(MemberRole.CREATOR)) {
+            throw new IllegalArgumentException("CREATOR만 수정, 삭제할 수 있음.");
+        }
 
         commentRepository.deleteById(id);
-
     }
 }
